Issue #797 Refactor message so it appears above service search result but not organization search results
