+++ src/components/search/search_component.tsx
@@ -2,20 +2,26 @@ import React, { useState } from 'react';
 import { InstantSearch, Index, connectSearchBox, connectInfiniteHits, connectConfigure } from 'react-instantsearch-native';
 import { SearchInputComponent } from './search_input_component';
 import { InfiniteHitsComponent } from './infinite_hits_component';
-import { colors } from '../../application/styles';
-import { Content } from 'native-base';
+import { colors, textStyles } from '../../application/styles';
+import { Content, Text } from 'native-base';
+import { View } from 'react-native';
 import { emptyComponent } from '../empty_component/empty_component';
 import { Locale } from '../../locale';
 import { LatLong } from './types';
 import { useFetchLatLongFromLocation } from './api/use_fetch_lat_long_from_location';
 import { toServiceSearchConfiguration } from './api/configuration';
 import { useTraceUpdate as useTraceComponentUpdates } from '../../helpers/debug';
+import { LinkTypes, AnalyticsLink } from '../link/link';
+import { buildLinkContext } from '../../sagas/analytics/events';
 import { ALGOLIA_SERVICES_INDEX, ALGOLIA_ORGANIZATIONS_INDEX } from 'react-native-dotenv';
+import { Trans } from '@lingui/react';
+import { SearchListSeparator } from './separators';
 
 export interface SearchComponentProps {
     readonly apiKey: string;
     readonly appId: string;
     readonly currentLocale: Locale;
+    readonly currentPath: string;
 }
 
 export interface SearchComponentActions {
@@ -41,12 +47,34 @@ export const SearchComponent: React.StatelessComponent<SearchComponentProps> = (
         <InstantSearch indexName={servicesIndex()} {...props} >
             <SearchInputConnectedComponent location={location} setLocation={setLocation} latLong={latLong} />
             <ConfigureConnectedComponent {...toServiceSearchConfiguration(latLong)} />
+            <InformationOnlyInEnglishNotice {...props} />
             <InfiniteHitsConnectedComponent />
             <Index indexName={organizationsIndex()}>
                 <InfiniteHitsConnectedComponent />
             </Index>
         </InstantSearch>
     </Content>;
+};
+
+const InformationOnlyInEnglishNotice = (props: Props): JSX.Element => {
+    const phoneNumber: string = '211';
+    const href = `tel: ${phoneNumber}`;
+    const linkContext = buildLinkContext('Service', 'BC-211');
+
+    return <View style={{ backgroundColor: colors.white }}>
+        <Text style={[textStyles.paragraphStyle,
+        {
+            marginTop: 20,
+            marginBottom: 20,
+            marginLeft: 20,
+            marginRight: 20,
+        }]}>
+            <Trans>Information about services is currently only available in English. For information in other languages, <AnalyticsLink
+                href={href} currentPath={props.currentPath} linkContext={linkContext} linkType={LinkTypes.phone} style={{ color: colors.teal }}
+            >please call BC211</AnalyticsLink>.</Trans>
+        </Text>
+        <SearchListSeparator />
+    </View>;
 };
 
 const servicesIndex = (): string => (
