+++ src/components/search/search_connected_component.ts
@@ -4,10 +4,15 @@ import { SearchComponentProps, SearchComponent, SearchComponentActions } from '.
 import { ALGOLIA_SEARCH_API_KEY } from 'react-native-dotenv';
 import { selectLocale } from '../../selectors/locale/select_locale';
 
-const mapStateToProps = (store: Store): SearchComponentProps => ({
+type OwnProps = {
+    readonly location: Location;
+};
+
+const mapStateToProps = (store: Store, ownProps: OwnProps): SearchComponentProps => ({
     apiKey: ALGOLIA_SEARCH_API_KEY,
     appId: 'MMYH1Z0D3O',
     currentLocale: selectLocale(store),
+    currentPath: ownProps.location.pathname,
 });
 
 const mapDispatchToProps = (_: Dispatch<void>): SearchComponentActions => ({
